Fix Java/C++ parity: renumberer range checks, warnings, widget formats
- Add missing isAreaVnum range checks for item/mobile/room vnums
- Match Java warning format: owner type, vnum, program#, line, offset
- Show warning count in RenumberWarningsWidget status bar
- Fix NewResetWidget combo to show name with value as item data

Fix parity: show "p " prefix in trigger field like Java

ProgramsWidget now displays the whole phrase "p " prefix directly
in the trigger edit field instead of hiding it behind the checkbox.
Matches the original Java UI behavior. Also remove redundant
headCanChange_ assignment in clearAreaData.

Fix type width mismatches: match Java long/BigInteger with qint64

Widen all fields that are long or BigInteger in the Java JAXB model
from int to qint64 in C++. This prevents potential data loss for
credits, cost, weight, economy values, vnums, reset args, and keeper
fields. Updated XML I/O (readLong/writeLong), Renumberer, MainWindow,
and all downstream widgets to use the wider types consistently.

Fix Java parity: FlagsWidget NONE logic, reset right-click, status timeouts, and more

- Fix FlagsWidget NONE checkbox: only checked when value is 0, checking it clears all flags
- Add ResetVnumEventFilter for right-click vnum override on reset combo boxes
- Add 5-second timeout to all statusBar messages
- Fix shop/repair delete to select previous item instead of jumping to first
- Remove dead areaCreated() method
- ProgramsWidget: track modified state, disable accept until changed
- Add type combo change detection in ProgramsWidget
- VnumQuestionWidget: widen signal from int to qint64
- Parse exitGridColumns from exits.xml, pass to NewExitWidget
- Fix exit button minimum size (40x30 -> 25x25) to match Java
- Always write nightdesc element (even when empty) to match Java
- Remove maxWidth constraints on mobile flag buttons in .ui
- Fix truncated "Edit Susceptib..." button label

Fix button sizing, FlagsWidget spacing, tray quit, and Mobiles layout

- Remove maximumSize constraints from all 22 QPushButtons to prevent text truncation
- Set sizePolicy hsizetype=Maximum on all flag/action buttons so they fit text without stretching
- Fix FlagsWidget.ui: remove fixed 70px label size and 0-margin overrides for cleaner Value: row
- Fix tray Quit: call QApplication::quit() in closeEvent so app terminates on macOS
- Add expanding spacers to Act Flags, Affected Flags, Credits rows on Mobiles tab to prevent justified layout

Fix BalloonWidget tooltip crash and square window on macOS

Clear lastToolTip_ in closeEvent to prevent use-after-free crash when
showing tooltip a second time. Replace Qt::ToolTip window type with
frameless bypass-WM window and WA_TranslucentBackground to eliminate
the rectangular window decoration macOS renders around tooltip windows.
Paint balloon shape onto transparent background with clip region instead
of relying on setMask which doesn't work reliably on macOS.

+++ src/gui/NewShopWidget.h
@@ -17,7 +17,7 @@ public:
     ~NewShopWidget();
 
 signals:
-    void vnumChosen(int vnum);
+    void vnumChosen(qint64 vnum);
 
 private slots:
     void onAcceptClicked();
